requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==3.7.1
api==0.0.0
attrs==23.1.0
bcrypt==4.1.2
certifi==2023.11.17
cffi==1.16.0
click==8.1.7
colorama==0.4.6
coverage==7.3.3
cryptography==41.0.7
db==0.0.0
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.105.0
greenlet==3.0.2
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.2
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
orjson==3.9.10
outcome==1.3.0.post0
packaging==23.2
passlib==1.7.4
pluggy==1.3.0
psycopg2==2.9.9
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.2
pydantic-extra-types==2.2.0
pydantic-settings==2.1.0
pydantic_core==2.14.5
PySocks==1.7.1
pytest==7.4.3
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
rsa==4.9
schemas==0.0.0
selenium==4.16.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.23
starlette==0.27.0
tomli==2.0.1
trio==0.23.2
trio-websocket==0.11.1
typing_extensions==4.9.0
ujson==5.9.0
unicorn==2.0.1.post1
urllib3==2.1.0
user-manager==0.0.0
utils==0.0.0
uvicorn==0.24.0.post1
watchfiles==0.21.0
websockets==12.0
wsproto==1.2.0
